Eliminate interface for LoadReportClient.
